+++ licenses/license-management.go
@@ -1,4 +1,4 @@
-package license
+package licenses
 
 import (
 	"context"
